+++ integrations/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v2.2.0] integrations
 ### Releases
 - [NPM v2.2.0] https://www.npmjs.com/package/@wingspan/integrations/v/2.2.0 - integrations
+
+## 2023-10-30 01:26:34
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/650da31d2b30f6000ce71dd7
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.3.0] integrations
+### Releases
+- [NPM v2.3.0] https://www.npmjs.com/package/@wingspan/integrations/v/2.3.0 - integrations